削除機能の修正

+++ microhr/views/company.py
@@ -6,6 +6,8 @@ from microhr.models import Work
 from microhr.forms import WorkForm
 from microhr.decorators import company_required
 from logging import getLogger
+
+from microhr.views.root import home
 logger = getLogger(__name__)
 
 
@@ -54,7 +56,10 @@ def work_edit(request, work_id):
 @login_required
 @company_required
 def work_delete(request, work_id):
-    """求人を削除する（未実装）"""
-    work = get_object_or_404(Work, pk=work_id)
-    work.delete()
-    return redirect('/')
+    """求人を削除する"""
+    if request.method == 'POST':
+        work = get_object_or_404(Work, pk=work_id)
+        work.delete()
+        return redirect(home)
+    else:
+        return HttpResponse("GETでの削除はできません、前の画面に戻ってください。")
